+++ generator_impl_test.go
@@ -10,17 +10,17 @@ const (
 )
 
 // Generator creator function.
-type generatorCreator func(workerId, datacenterId uint64) (Generator, error)
+type generatorCreator func(workerId, datacenterId int) (Generator, error)
 
 // Test constants.
 func TestConstants(t *testing.T) {
-	if sequenceMask != uint64(4095) {
+	if sequenceMask != int64(4095) {
 		t.Errorf("Expected sequence mask to be 4095, but it is %d", sequenceMask)
 	}
 }
 
 // Test generator implementation.
-func testGeneratorImplementation(t *testing.T, gName string, creator generatorCreator, epoch uint64) {
+func testGeneratorImplementation(t *testing.T, gName string, creator generatorCreator, epoch int64) {
 	// Creating new generator with out of range IDs returns expected error.
 	g, err := creator(32, 0)
 	if err != ErrWorkerIdOutOfRange {
@@ -41,8 +41,8 @@ func testGeneratorImplementation(t *testing.T, gName string, creator generatorCr
 	// Creating new generator with valid IDs returns a generator and no
 	// error.
 	for _, args := range []struct {
-		workerId     uint64
-		datacenterId uint64
+		workerId     int
+		datacenterId int
 	}{
 		{0, 0},
 		{15, 16},
@@ -77,7 +77,7 @@ func testGeneratorImplementation(t *testing.T, gName string, creator generatorCr
 		t.Fatalf("Unexpected error creating generator: %v", err)
 	}
 
-	now := uint64(time.Now().UTC().UnixNano() / 1000000)
+	now := int64(time.Now().UTC().UnixNano() / 1000000)
 	timestamp := g.Timestamp()
 
 	if timestamp < now || timestamp > now+acceptedClockDriftMs {
@@ -87,10 +87,10 @@ func testGeneratorImplementation(t *testing.T, gName string, creator generatorCr
 	// A generator returns unique IDs - we'll prove this by generating
 	// 10^7 of them.
 	count := 10000000
-	generatedIds := make(map[uint64]struct{}, count)
+	generatedIds := make(map[int64]struct{}, count)
 
 	for c := 0; c < count; c++ {
-		now = uint64(time.Now().UTC().UnixNano()/1000000) - g.Epoch()
+		now = int64(time.Now().UTC().UnixNano()/1000000) - g.Epoch()
 
 		id, err := g.NextId()
 		if err != nil {
@@ -136,7 +136,7 @@ func TestGenerator(t *testing.T) {
 
 // Test generator with custom epoch.
 func TestGeneratorEpoch(t *testing.T) {
-	testGeneratorImplementation(t, "NewGeneratorEpoch", func(w, d uint64) (Generator, error) {
+	testGeneratorImplementation(t, "NewGeneratorEpoch", func(w, d int) (Generator, error) {
 		return NewGeneratorEpoch(w, d, 1400027069000)
 	}, 1400027069000)
 }
